project/requirements.txt
fastapi
uvicorn
asyncpg
tortoise-orm
pytest
requests
gunicorn
uvloop
httptools
pytest-cov
flake8
black
isort